+++ src/main/java/com/codelion/animalcare/domain/doctorqna/dto/request/AnswerSaveRequestDto.java
@@ -1,7 +1,7 @@
 package com.codelion.animalcare.domain.doctorqna.dto.request;
 
-import com.codelion.animalcare.domain.doctorqna.repository.Answer;
-import com.codelion.animalcare.domain.doctorqna.repository.Question;
+import com.codelion.animalcare.domain.doctorqna.entity.Answer;
+import com.codelion.animalcare.domain.doctorqna.entity.Question;
 import com.codelion.animalcare.domain.user.entity.Doctor;
 import lombok.Builder;
 import lombok.Getter;
