+++ src/main/java/com/codelion/animalcare/domain/doctorqna/dto/response/AnswerResponseDto.java
@@ -1,7 +1,7 @@
 package com.codelion.animalcare.domain.doctorqna.dto.response;
 
-import com.codelion.animalcare.domain.doctorqna.repository.Answer;
-import com.codelion.animalcare.domain.doctorqna.repository.Question;
+import com.codelion.animalcare.domain.doctorqna.entity.Answer;
+import com.codelion.animalcare.domain.doctorqna.entity.Question;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
