+++ src/main/java/com/codelion/animalcare/domain/doctorqna/repository/AnswerRepository.java
@@ -1,5 +1,6 @@
 package com.codelion.animalcare.domain.doctorqna.repository;
 
+import com.codelion.animalcare.domain.doctorqna.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AnswerRepository extends JpaRepository<Answer, Long> {
