+++ src/main/java/com/codelion/animalcare/domain/doctorqna/repository/QuestionLikeRepository.java
@@ -1,5 +1,6 @@
 package com.codelion.animalcare.domain.doctorqna.repository;
 
+import com.codelion.animalcare.domain.doctorqna.entity.QuestionLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface QuestionLikeRepository extends JpaRepository<QuestionLike, Long> {
