+++ src/main/java/com/codelion/animalcare/domain/doctorqna/repository/QuestionRepository.java
@@ -1,5 +1,6 @@
 package com.codelion.animalcare.domain.doctorqna.repository;
 
+import com.codelion.animalcare.domain.doctorqna.entity.Question;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
