+++ src/main/java/com/codelion/animalcare/domain/doctorqna/service/QuestionCommandService.java
@@ -2,8 +2,8 @@ package com.codelion.animalcare.domain.doctorqna.service;
 
 import com.codelion.animalcare.domain.doctorqna.dto.request.QuestionSaveRequestDto;
 import com.codelion.animalcare.domain.doctorqna.dto.request.QuestionUpdateRequestDto;
-import com.codelion.animalcare.domain.doctorqna.repository.Question;
-import com.codelion.animalcare.domain.doctorqna.repository.QuestionLike;
+import com.codelion.animalcare.domain.doctorqna.entity.Question;
+import com.codelion.animalcare.domain.doctorqna.entity.QuestionLike;
 import com.codelion.animalcare.domain.doctorqna.repository.QuestionLikeRepository;
 import com.codelion.animalcare.domain.doctorqna.repository.QuestionRepository;
 import com.codelion.animalcare.domain.user.entity.Member;
@@ -20,10 +20,11 @@ import java.security.Principal;
 @Transactional
 public class QuestionCommandService {
 
-    private final QuestionRepository questionRepository;
-    private final MemberService memberService;
-    private final QuestionLikeRepository questionLikeRepository;
     private final QuestionQueryService questionQueryService;
+    private final QuestionRepository questionRepository;
+    private final QuestionLikeRepository questionLikeRepository;
+    private final MemberService memberService;
+
 
 
     public Long save(QuestionSaveRequestDto questionSaveRequestDto, Principal principal) {
